Fix: userIdx to userToken & catch exception

+++ src/main/java/com/weave/weaveserver/service/TeamService.java
@@ -2,17 +2,23 @@
 package com.weave.weaveserver.service;
 
 import com.weave.weaveserver.config.exception.BadRequestException;
+import com.weave.weaveserver.config.exception.NotFoundException;
+import com.weave.weaveserver.config.jwt.TokenService;
 import com.weave.weaveserver.domain.Belong;
 import com.weave.weaveserver.domain.Team;
 import com.weave.weaveserver.domain.User;
+import com.weave.weaveserver.dto.JsonResponse;
 import com.weave.weaveserver.dto.TeamRequest;
 import com.weave.weaveserver.dto.TeamResponse;
 import com.weave.weaveserver.repository.BelongRepository;
 import com.weave.weaveserver.repository.TeamRepository;
 import com.weave.weaveserver.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -21,13 +27,19 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class TeamService {
 
+    @Autowired
+    private TokenService tokenService;
+
     private final TeamRepository teamRepository;
     private final UserRepository userRepository;
     private final BelongRepository belongRepository;
 
     @Transactional
-    public void createTeam(Long leaderIdx, TeamRequest.createReq req) {
-        User leader = userRepository.getReferenceById(leaderIdx);
+    public void createTeam(TeamRequest.createReq req, HttpServletRequest httpServletRequest) {
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User leader = userRepository.findUserByEmail(userEmail);
+
+        System.out.println(leader.getUserIdx());
 
         Team team = Team.builder()
                 .leader(leader)
@@ -47,20 +59,35 @@ public class TeamService {
     }
 
     @Transactional
-    public int addMember(Long teamIdx, TeamRequest.addMemberReq req) {
+    public int addMember(Long teamIdx, TeamRequest.addMemberReq req, HttpServletRequest httpServletRequest) {
 
-        // team creator == request user (생성하려는 사용자와 팀짱이 동일인인지 확인)
-        Team team = teamRepository.getReferenceById(teamIdx);
-//        User creator = team.getLeader();
-//
-//        if(!(creator.getUserIdx().equals(req.getLeaderIdx()))){
-//            System.out.println("생성자와 팀짱이 일치하지 않음");
-//            return 0;
-//        }
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
+        User leader = team.getLeader();
+
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User creator = userRepository.findUserByEmail(userEmail);
+
+        if(leader.equals(creator)){
+            //생성하려는 사용자와 팀짱이 동일
+            System.out.println("same");
+
+
+        } else {
+            //생성하려는 사용자와 팀짱이 동일하지 않음
+            System.out.println("not same");
+            return 3000;
+        }
+
 
         // find user (존재하는 사용자인지 확인)
         User invitedUser = userRepository.findUserByEmail(req.getEmail());
         if(invitedUser != null){
+            // 이미 팀에 존재하는 유저인지 확인
+            //System.out.println("이미 초대된 팀원입니다");
+
             System.out.println("초대 팀원 존재 : "+ invitedUser.getEmail());
 
             Long userIdx = invitedUser.getUserIdx();
@@ -87,7 +114,12 @@ public class TeamService {
 
     @Transactional
     public List<TeamResponse.getMemberList> getMembers(Long teamIdx){
-        List<User> userList = belongRepository.findAllByTeamIdx(teamIdx);
+
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
+        List<User> userList = belongRepository.findAllByTeamIdx(team.getTeamIdx());
 
         List<TeamResponse.getMemberList> memberList = userList.stream().map(user -> new TeamResponse.getMemberList(
                 user.getUserIdx(),
@@ -99,8 +131,13 @@ public class TeamService {
     }
 
     @Transactional
-    public List<TeamResponse.teamResponse> getMyTeams(Long userIdx){
-        List<Team> teams = belongRepository.findAllByUserIdx(userIdx);
+    public List<TeamResponse.teamResponse> getMyTeams(HttpServletRequest httpServletRequest){
+
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User user = userRepository.findUserByEmail(userEmail);
+        //System.out.println(userEmail);
+        //System.out.println(user.getUserIdx());
+        List<Team> teams = belongRepository.findAllByUserIdx(user.getUserIdx());
 
         List<TeamResponse.teamResponse> teamList = teams.stream().map(team -> new TeamResponse.teamResponse(
                 team.getTeamIdx(),
@@ -114,25 +151,87 @@ public class TeamService {
     }
 
     @Transactional
-    public void deleteTeam(Long teamIdx, TeamRequest.deleteTeamReq req){
+    public void deleteTeam(Long teamIdx, HttpServletRequest httpServletRequest){
         //팀짱인지 확인 과정 필요
 
-        teamRepository.deleteByTeamIdx(teamIdx);
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
+        User leader = team.getLeader();
+
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User requester = userRepository.findUserByEmail(userEmail);
+
+        if(leader.equals(requester)){
+            //요청 사용자와 팀짱이 동일
+            System.out.println("same");
+            teamRepository.deleteByTeamIdx(teamIdx);
+
+        } else {
+            //요청 사용자와 팀짱이 동일하지 않음
+            System.out.println("not same");
+        }
+
     }
 
     @Transactional
-    public void deleteMember(Long teamIdx, TeamRequest.deleteMemberReq req){
-        //팀짱인지 확인 과정 필요
+    public void deleteMember(Long teamIdx, TeamRequest.deleteMemberReq req, HttpServletRequest httpServletRequest){
 
-        Belong deleteUser = belongRepository.findUserByIndex(teamIdx, req.getUserIdx());
-        belongRepository.deleteById(deleteUser.getBelongIdx());
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
+        User leader = team.getLeader();
+
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User requester = userRepository.findUserByEmail(userEmail);
+
+        if(leader.equals(requester)){
+            //요청 사용자와 팀짱이 동일
+            System.out.println("same");
+
+
+
+            // 삭제하려는 팀원이 팀 소속인지 확인
+
+
+
+            // 소속이라면
+            Belong deleteUser = belongRepository.findUserByIndex(teamIdx, req.getUserIdx());
+            belongRepository.deleteById(deleteUser.getBelongIdx());
+
+        } else {
+            //요청 사용자와 팀짱이 동일하지 않음
+            System.out.println("not same");
+        }
+
+
     }
 
     @Transactional
-    public void updateTeam(Long teamIdx, TeamRequest.updateTeamReq req){
+    public void updateTeam(Long teamIdx, TeamRequest.updateTeamReq req, HttpServletRequest httpServletRequest){
         //팀짱인지 확인 필요
 
-        Team team = teamRepository.getReferenceById(teamIdx);
-        team.updateTeam(req.getTitle(), req.getStartDate(), req.getEndDate(), req.getImgUrl());
+        // 해당 팀 IDX 가 존재하지 않는 경우에 대한 예외 처리
+        Team team = teamRepository.findById(teamIdx)
+                .orElseThrow(() -> new NotFoundException("해당 팀이 존재하지 않습니다."));
+
+        User leader = team.getLeader();
+
+        String userEmail = tokenService.getUserEmail(httpServletRequest);
+        User requester = userRepository.findUserByEmail(userEmail);
+
+        if(leader.equals(requester)){
+            //요청 사용자와 팀짱이 동일
+            System.out.println("same");
+
+            //업데이트
+            team.updateTeam(req.getTitle(), req.getStartDate(), req.getEndDate(), req.getImgUrl());
+
+        } else {
+            //요청 사용자와 팀짱이 동일하지 않음
+            System.out.println("not same");
+        }
     }
 }
